pyproject.toml
[project]
name = "failure-first-planner"
version = "1.0.0"
description = "Failure-First Event Planning Agent for live music events"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "geopy>=2.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that call the real OpenAI API (deselect with '-m not integration')",
]
testpaths = ["tests"]
